Split index into executable and logic file
